requirements.txt
numpy>=1.26
pandas>=2.1
scipy>=1.11
cvxpy>=1.4
osqp>=0.6
ecos>=2.0
statsmodels>=0.14
matplotlib>=3.8
pyyaml>=6.0
pyarrow>=15.0
